+++ polaris-react/src/components/Navigation/Navigation.stories.tsx
@@ -31,18 +31,20 @@ export function Default() {
         <Navigation.Section
           items={[
             {
-              url: '/',
+              url: '#',
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               badge: '15',
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
             },
@@ -78,62 +80,70 @@ export function WithMultipleSecondaryNavigations() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               badge: '15',
               subNavigationItems: [
                 {
-                  url: '/admin/orders/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   selected: false,
                   label: 'Collections',
                 },
                 {
-                  url: '/admin/orders/inventory',
+                  url: '#',
                   disabled: false,
                   label: 'Inventory',
+                  excludePaths: ['#'],
                 },
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Marketing',
               icon: MarketingMinor,
               badge: '15',
               subNavigationItems: [
                 {
-                  url: '/admin/analytics/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   selected: false,
                   label: 'Reports',
                 },
                 {
-                  url: '/admin/analytics/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Live view',
                 },
               ],
             },
             {
-              url: '/admin/products',
+              url: '#',
               label: 'Products',
               icon: ProductsMinor,
               selected: true,
               subNavigationItems: [
                 {
-                  url: '/?path=/story/all-components-navigation--navigation-with-multiple-secondary-navigations',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   selected: false,
                   label: 'Collections',
                 },
                 {
-                  url: '/admin/products/inventory',
+                  url: '#',
                   disabled: false,
                   selected: true,
                   label: 'Inventory',
@@ -155,30 +165,34 @@ export function WithAnActiveRootItemWithSecondaryNavigationItems() {
           duplicateRootItem
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               badge: '15',
             },
             {
-              url: '/admin/products',
+              url: '#',
               label: 'Products',
               icon: ProductsMinor,
               selected: true,
               subNavigationItems: [
                 {
-                  url: '/admin/products/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   selected: false,
                   label: 'Collections',
                 },
                 {
-                  url: '/admin/products/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Inventory',
                 },
@@ -198,17 +212,20 @@ export function WithASecondaryActionForASectionAndASectionTitle() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
             },
@@ -218,7 +235,8 @@ export function WithASecondaryActionForASectionAndASectionTitle() {
           title="Sales channels"
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Online Store',
               icon: OnlineStoreMinor,
             },
@@ -241,16 +259,18 @@ export function WithASecondaryActionForAnItem() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               secondaryAction: {
-                url: '/admin/orders/add',
+                url: '#',
                 accessibilityLabel: 'Add an order',
                 icon: CirclePlusOutlineMinor,
                 tooltip: {
@@ -259,7 +279,8 @@ export function WithASecondaryActionForAnItem() {
               },
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
             },
@@ -277,24 +298,27 @@ export function WithMultipleSecondaryActionsForAnItem() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               badge: '123',
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
               badge: '2',
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -304,13 +328,14 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Long multi-line label',
               icon: PlusMinor,
               badge: '125',
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -328,13 +353,14 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Long truncated single-line text label',
               icon: PlusMinor,
               truncateText: true,
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -354,22 +380,22 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
               subNavigationItems: [
                 {
-                  url: '/path/to/place/index',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'New item',
                 },
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Floating actions on multi-line text label',
               icon: PlusMinor,
               badge: '15',
-              selected: true,
               displayActionsOnHover: true,
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -389,20 +415,22 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
               subNavigationItems: [
                 {
-                  url: '/admin/products/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
-                  selected: false,
                   label: 'Collections',
                 },
                 {
-                  url: '/admin/products/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Inventory',
                 },
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Floating actions on truncated single-line text label',
               icon: PlusMinor,
               badge: '15',
@@ -410,7 +438,7 @@ export function WithMultipleSecondaryActionsForAnItem() {
               displayActionsOnHover: true,
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -429,13 +457,14 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'One floating action on multi-line text label',
               icon: PlusMinor,
               displayActionsOnHover: true,
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -445,7 +474,8 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'One floating action on truncated single-line text label',
               icon: PlusMinor,
               badge: '15',
@@ -453,7 +483,7 @@ export function WithMultipleSecondaryActionsForAnItem() {
               displayActionsOnHover: true,
               secondaryActions: [
                 {
-                  url: '/admin/products/add',
+                  url: '#',
                   accessibilityLabel: 'Add a product',
                   icon: CirclePlusOutlineMinor,
                   tooltip: {
@@ -463,14 +493,14 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Marketing',
               icon: MarketingMinor,
               badge: '15',
               selected: true,
               secondaryActions: [
                 {
-                  url: '/path/to/place/view',
+                  url: '#',
                   accessibilityLabel: 'View campaign',
                   icon: ViewMinor,
                   tooltip: {
@@ -480,7 +510,8 @@ export function WithMultipleSecondaryActionsForAnItem() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Logout',
               icon: LogOutMinor,
             },
@@ -498,26 +529,26 @@ export function WithTruncationForVariousStates() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'A very long label to ellipsize',
               truncateText: true,
               icon: OrdersMinor,
               selected: false,
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Not truncated',
               icon: OrdersMinor,
               selected: false,
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Lengthy label with secondary action',
               icon: OrdersMinor,
               selected: false,
               truncateText: true,
               secondaryAction: {
-                url: '/admin/orders/add',
+                url: '#',
                 accessibilityLabel: 'Add an order',
                 icon: CirclePlusOutlineMinor,
                 tooltip: {
@@ -526,12 +557,12 @@ export function WithTruncationForVariousStates() {
               },
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Lengthy non-truncated label with secondary action',
               icon: OrdersMinor,
               selected: false,
               secondaryAction: {
-                url: '/admin/orders/add',
+                url: '#',
                 accessibilityLabel: 'Add an order',
                 icon: CirclePlusOutlineMinor,
                 tooltip: {
@@ -540,21 +571,22 @@ export function WithTruncationForVariousStates() {
               },
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Lengthy label with badge',
               truncateText: true,
               badge: 'Old',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Lengthy label with secondary action',
               icon: OrdersMinor,
               selected: false,
               truncateText: true,
               badge: 'Old',
               secondaryAction: {
-                url: '/admin/orders/add',
+                url: '#',
                 accessibilityLabel: 'Add an order',
                 icon: CirclePlusOutlineMinor,
                 tooltip: {
@@ -563,19 +595,21 @@ export function WithTruncationForVariousStates() {
               },
             },
             {
-              url: '/admin/products',
+              url: '#',
               label: 'Truncated secondary navigation items',
               icon: ProductsMinor,
               selected: true,
               truncateText: true,
               subNavigationItems: [
                 {
-                  url: '/admin/products/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Something longer than collections',
                 },
                 {
-                  url: '/admin/products/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Inventoy',
                 },
@@ -595,17 +629,20 @@ export function WithSectionRollup() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
             },
@@ -629,17 +666,20 @@ export function WithSectionSeparator() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
             },
@@ -648,7 +688,8 @@ export function WithSectionSeparator() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Online Store',
               icon: OnlineStoreMinor,
             },
@@ -667,17 +708,20 @@ export function WithVariousStatesAndSecondaryElements() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Inactive item',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Item with indicator',
               icon: HomeMinor,
               subNavigationItems: [
                 {
-                  url: '/path/to/place/index',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   new: true,
                   label: 'New item',
@@ -685,24 +729,26 @@ export function WithVariousStatesAndSecondaryElements() {
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'New item',
               new: true,
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Badged item',
               badge: 'Old',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Active with secondary action',
               icon: OrdersMinor,
               selected: true,
               secondaryAction: {
-                url: '/admin/orders/add',
+                url: '#',
                 accessibilityLabel: 'Add an order',
                 icon: CirclePlusOutlineMinor,
                 tooltip: {
@@ -711,37 +757,40 @@ export function WithVariousStatesAndSecondaryElements() {
               },
             },
             {
-              url: '/admin/products',
+              url: window.location.href,
               label: 'Active item with sub navigation',
               icon: ProductsMinor,
               selected: true,
               subNavigationItems: [
                 {
-                  url: '/admin/products',
-                  disabled: false,
+                  url: window.location.href,
                   selected: true,
+                  disabled: false,
                   label: 'Selected sub item',
                 },
                 {
-                  url: '/admin/products/transfers',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Default sub item',
                 },
                 {
-                  url: '/admin/products/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: true,
                   label: 'Disabled sub item',
                 },
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Disabled item',
               icon: CustomersMinor,
               disabled: true,
             },
             {
-              url: '/path/to/place',
+              url: '#',
               label: 'Overflow item',
               icon: MarketingMinor,
             },
@@ -757,26 +806,29 @@ export function WithVariousStatesAndSecondaryElements() {
           title="These icons should have the same color"
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Icon as svg',
               icon: OnlineStoreMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Icon as img',
               icon: '<svg viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M1.791 2.253l-.597 3.583A1 1 0 0 0 2.18 7h.893a1.5 1.5 0 0 0 1.342-.83L5 5l.585 1.17A1.5 1.5 0 0 0 6.927 7h1.146a1.5 1.5 0 0 0 1.342-.83L10 5l.585 1.17a1.5 1.5 0 0 0 1.342.83h1.146a1.5 1.5 0 0 0 1.342-.83L15 5l.585 1.17a1.5 1.5 0 0 0 1.342.83h.893a1 1 0 0 0 .986-1.164l-.597-3.583A1.5 1.5 0 0 0 16.729 1H3.271a1.5 1.5 0 0 0-1.48 1.253zM4 18.5A1.5 1.5 0 0 1 5.5 17H8v-3h4v3h2.5a1.5 1.5 0 0 1 1.5 1.5v.5H4v-.5z"></path><path d="M2 9h2v4h12V9h2v4.5a1.5 1.5 0 0 1-1.5 1.5h-13A1.5 1.5 0 0 1 2 13.5V9z"></path></svg>',
             },
             {
-              url: '/',
+              url: '#',
               label: 'Icon as img – Active',
               icon: '<svg viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M1.791 2.253l-.597 3.583A1 1 0 0 0 2.18 7h.893a1.5 1.5 0 0 0 1.342-.83L5 5l.585 1.17A1.5 1.5 0 0 0 6.927 7h1.146a1.5 1.5 0 0 0 1.342-.83L10 5l.585 1.17a1.5 1.5 0 0 0 1.342.83h1.146a1.5 1.5 0 0 0 1.342-.83L15 5l.585 1.17a1.5 1.5 0 0 0 1.342.83h.893a1 1 0 0 0 .986-1.164l-.597-3.583A1.5 1.5 0 0 0 16.729 1H3.271a1.5 1.5 0 0 0-1.48 1.253zM4 18.5A1.5 1.5 0 0 1 5.5 17H8v-3h4v3h2.5a1.5 1.5 0 0 1 1.5 1.5v.5H4v-.5z"></path><path d="M2 9h2v4h12V9h2v4.5a1.5 1.5 0 0 1-1.5 1.5h-13A1.5 1.5 0 0 1 2 13.5V9z"></path></svg>',
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Other secondary action',
               icon: '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16"><path d="M2.771 14.055A8 8 0 1 1 16 8c0 2.385-1.086 4.629-2.704 5.983A7.956 7.956 0 0 1 7.991 16c-.613 0-2.022-.003-5-.01h-.928l.708-1.935zm5.973-6.757c-.554-.302-.837-.565-.837-.92 0-.446.4-.735 1.017-.735a3.62 3.62 0 0 1 1.365.302l.502-1.577S10.328 4 8.963 4C7.057 4 5.73 5.117 5.73 6.68c0 .894.618 1.565 1.442 2.05.67.382.901.658.901 1.065 0 .42-.334.762-.952.762-.915 0-1.79-.486-1.79-.486l-.54 1.577s.797.552 2.15.552c1.956 0 3.373-.986 3.373-2.76-.013-.959-.721-1.642-1.571-2.142z"/></svg>',
               secondaryAction: {
-                url: '/path/to/place/view',
+                url: '#',
                 accessibilityLabel: 'View your online store',
                 icon: ViewMinor,
                 tooltip: {
@@ -807,18 +859,21 @@ export function WithAriaLabelledby() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               badge: '15',
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMinor,
             },
@@ -836,23 +891,27 @@ export function UsingMajorIcons() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMajor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMajor,
               badge: '15',
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Products',
               icon: ProductsMajor,
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Customers',
               icon: CustomersMajor,
             },
@@ -870,11 +929,12 @@ export function WithBadgeAndSecondaryAction() {
         <Navigation.Section
           items={[
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Home',
               icon: HomeMinor,
               secondaryAction: {
-                url: '/path/to/place/view',
+                url: '#',
                 accessibilityLabel: 'View your online store',
                 icon: ViewMinor,
                 tooltip: {
@@ -883,31 +943,34 @@ export function WithBadgeAndSecondaryAction() {
               },
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Orders',
               icon: OrdersMinor,
               badge: '15',
               subNavigationItems: [
                 {
-                  url: '/admin/orders/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
-                  selected: false,
                   label: 'Collections',
                 },
                 {
-                  url: '/admin/orders/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Inventory',
                 },
               ],
             },
             {
-              url: '/path/to/place',
+              url: '#',
+              excludePaths: ['#'],
               label: 'Marketing',
               icon: MarketingMinor,
               badge: '15',
               secondaryAction: {
-                url: '/path/to/place/view',
+                url: '#',
                 accessibilityLabel: 'View your online store',
                 icon: ViewMinor,
                 tooltip: {
@@ -916,34 +979,34 @@ export function WithBadgeAndSecondaryAction() {
               },
               subNavigationItems: [
                 {
-                  url: '/admin/analytics/collections',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
-                  selected: false,
                   label: 'Reports',
                 },
                 {
-                  url: '/admin/analytics/inventory',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
                   label: 'Live view',
                 },
               ],
             },
             {
-              url: '/admin/products',
+              url: '#',
               label: 'Products',
               icon: ProductsMinor,
               selected: true,
               subNavigationItems: [
                 {
-                  url: '/?path=/story/all-components-navigation--navigation-with-multiple-secondary-navigations',
+                  url: '#',
+                  excludePaths: ['#'],
                   disabled: false,
-                  selected: false,
                   label: 'Collections',
                 },
                 {
-                  url: '/admin/products/inventory',
+                  url: '#',
                   disabled: false,
-                  selected: true,
                   label: 'Inventory',
                 },
               ],