+++ polaris.shopify.com/pages/examples/index-table-with-all-of-its-elements.tsx
@@ -14,7 +14,7 @@ function IndexTableWithAllElementsExample() {
   const customers = [
     {
       id: '3417',
-      url: 'customers/341',
+      url: '#',
       name: 'Mae Jemison',
       location: 'Decatur, USA',
       orders: 20,
@@ -22,7 +22,7 @@ function IndexTableWithAllElementsExample() {
     },
     {
       id: '2567',
-      url: 'customers/256',
+      url: '#',
       name: 'Ellen Ochoa',
       location: 'Los Angeles, USA',
       orders: 30,