+++ polaris.shopify.com/pages/examples/page-with-custom-primary-action.tsx
@@ -5,7 +5,7 @@ import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
 function PageExample() {
   return (
     <Page
-      breadcrumbs={[{content: 'Settings', url: '/settings'}]}
+      breadcrumbs={[{content: 'Settings', url: '#'}]}
       title="General"
       primaryAction={
         <Button