+++ polaris.shopify.com/pages/examples/resource-list-with-bulk-actions.tsx
@@ -13,13 +13,13 @@ function ResourceListWithBulkActionsExample() {
   const items = [
     {
       id: 103,
-      url: 'customers/341',
+      url: '#',
       name: 'Mae Jemison',
       location: 'Decatur, USA',
     },
     {
       id: 203,
-      url: 'customers/256',
+      url: '#',
       name: 'Ellen Ochoa',
       location: 'Los Angeles, USA',
     },